+++ src/stories/SearchHome.stories.js
@@ -2,7 +2,7 @@ import { action } from "@storybook/addon-actions"
 import SearchHome from "@/lib-components/SearchHome"
 
 export default {
-    title: "Search / Home",
+    title: "Search / Search Home",
     component: SearchHome,
 }